+++ code/pali_stemmer.py
