+++ code/pali_vectorize_all.py
@@ -1,5 +1,5 @@
-from vectorizer import Vectorizer
-from file_mngr import FileMngr
+from code.pali_vectorizer import Vectorizer
+from code.pali_file_mngr import FileMngr
 
 import multiprocessing
 import os
